Remove meltpool example, remove unused postprocessors.
